Components: Move G2 Flex into UI
