+++ docs/manifest.json
@@ -924,12 +924,6 @@
 		"parent": "components"
 	},
 	{
-		"title": "Next",
-		"slug": "next",
-		"markdown_source": "../packages/components/src/flex/next/README.md",
-		"parent": "components"
-	},
-	{
 		"title": "Flex",
 		"slug": "flex",
 		"markdown_source": "../packages/components/src/flex/README.md",
