+++ packages/components/src/ui/flex/flex-block.js
@@ -1,7 +1,7 @@
 /**
  * Internal dependencies
  */
-import { createComponent } from '../../utils';
+import { createComponent } from '../utils';
 import { useFlexBlock } from './use-flex-block';
 
 /**
