+++ packages/components/src/ui/flex/flex-item.js
@@ -1,7 +1,7 @@
 /**
  * Internal dependencies
  */
-import { createComponent } from '../../utils';
+import { createComponent } from '../utils';
 import { useFlexItem } from './use-flex-item';
 
 /**
