+++ packages/components/src/ui/flex/flex.js
@@ -1,7 +1,7 @@
 /**
  * Internal dependencies
  */
-import { createComponent } from '../../utils';
+import { createComponent } from '../utils';
 import { useFlex } from './use-flex';
 
 /**
