+++ packages/components/src/ui/flex/index.js
@@ -1,0 +1,7 @@
+export { default as Flex } from './flex';
+export { default as FlexItem } from './flex-item';
+export { default as FlexBlock } from './flex-block';
+
+export * from './use-flex';
+export * from './use-flex-item';
+export * from './use-flex-block';
