+++ packages/components/src/ui/flex/test/index.js
@@ -6,7 +6,7 @@ import { render } from '@testing-library/react';
 /**
  * Internal dependencies
  */
-import View from '../../../view';
+import { View } from '../../view';
 import Flex from '../flex';
 import FlexItem from '../flex-item';
 import FlexBlock from '../flex-block';
